Move and rename Meny -> CurrentElement
