+++ src/InformasjonsVisning/Katalog/Relasjoner/Relasjon.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Navigeringsliste from "../../../Navigering/Navigeringsliste/Navigeringsliste";
+import Navigeringsliste from "../../../Kartlag/CurrentElement/Navigeringsliste/Navigeringsliste";
 import SectionExpand from "../../../GjenbruksElement/SectionExpand";
 import { titler } from "./Titler";
 const Relasjon = props => {
