+++ src/Kartlag/CurrentElement/Navigeringsliste/NavigeringslisteFunksjoner/kodeSuffix.js
