+++ src/Kartlag/CurrentElement/Navigeringsliste/VolumIndikator.js
